+++ test/login_test.exs
@@ -2,7 +2,7 @@ defmodule Openmaize.LoginTest do
   use ExUnit.Case
   use Plug.Test
 
-  alias Openmaize.{Login, LoginTools, QueryTools, TestRepo, User}
+  alias Openmaize.{DB, Login, LoginTools, TestRepo, User}
 
   setup_all do
     user = %{email: "ray@mail.com", username: "ray", role: "user", password: "hard2guess",
@@ -19,7 +19,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "login succeeds with username" do
-    opts = {true, :cookie, {0, 1440}, :username, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, :username, &DB.find_user/2}
     conn = call("ray", "hard2guess", "username", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/users"}
@@ -28,7 +28,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "login succeeds with email" do
-    opts = {true, :cookie, {0, 1440}, :email, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, :email, &DB.find_user/2}
     conn = call("ray@mail.com", "hard2guess", "email", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/users"}
@@ -37,7 +37,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "login fails for incorrect password" do
-    opts = {true, :cookie, {0, 1440}, :email, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, :email, &DB.find_user/2}
     conn = call("ray@mail.com", "oohwhatwasitagain", "email", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/login"}
@@ -46,7 +46,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "login fails for invalid email" do
-    opts = {true, :cookie, {0, 1440}, :email, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, :email, &DB.find_user/2}
     conn = call("dick@mail.com", "hard2guess", "email", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/login"}
@@ -55,7 +55,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "multiple possible unique ids - email for email_username func" do
-    opts = {true, :cookie, {0, 1440}, &LoginTools.email_username/1, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, &LoginTools.email_username/1, &DB.find_user/2}
     conn = call("ray@mail.com", "hard2guess", "email", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/users"}
@@ -64,7 +64,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "multiple possible unique ids - username for email_username func" do
-    opts = {true, :cookie, {0, 1440}, &LoginTools.email_username/1, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, &LoginTools.email_username/1, &DB.find_user/2}
     conn = call("ray", "hard2guess", "email", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/users"}
@@ -73,7 +73,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "multiple possible unique ids - phone for phone_username func" do
-    opts = {true, :cookie, {0, 1440}, &LoginTools.phone_username/1, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, &LoginTools.phone_username/1, &DB.find_user/2}
     conn = call("081555555", "hard2guess", "phone", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/users"}
@@ -82,7 +82,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "multiple possible unique ids - username for phone_username func" do
-    opts = {true, :cookie, {0, 1440}, &LoginTools.phone_username/1, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, &LoginTools.phone_username/1, &DB.find_user/2}
     conn = call("ray", "hard2guess", "phone", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/users"}
@@ -91,7 +91,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "fail login with multiple possible unique ids - phone for phone_username func" do
-    opts = {true, :cookie, {0, 1440}, &LoginTools.phone_username/1, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, &LoginTools.phone_username/1, &DB.find_user/2}
     conn = call("081555555", "oohwhatwasitagain", "phone", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/login"}
@@ -100,7 +100,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "fail login with multiple possible unique ids - username for phone_username func" do
-    opts = {true, :cookie, {0, 1440}, &LoginTools.phone_username/1, &QueryTools.find_user/2}
+    opts = {true, :cookie, {0, 1440}, &LoginTools.phone_username/1, &DB.find_user/2}
     conn = call("rav", "hard2guess", "phone", opts)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
       {"location", "/login"}